Rename `ares` dir to `core`
